requirements.txt
joblib
pandas
scikit-learn
nltk
textblob
python-dotenv
matplotlib
seaborn
contractions
numpy
scipy
textstat
gensim
wordcloud